requirements.txt
duckdb
streamlit
polars
huggingface_hub
datasets
python-dotenv